imported all the images and restructured the images-folder
